Add abstraction to cover Philosopher problem
