+++ src/main/kotlin/uzhnu/edu/tdspc/Philosopher.kt
@@ -1,0 +1,3 @@
+package uzhnu.edu.tdspc
+
+abstract class Philosopher : Runnable
